# Speed up sum by using reasonable read buffer sizes.