@sylvestre 2022-07-24
can you please document your benchmark with hyperfine in src/uu/sum/BENCHMARKING.md ?
thanks