@resistor 2022-07-25
> it broke some tests btw:
> 
> ```
> 
> ---- test_sum::test_sysv_stdin stdout ----
> current_directory_resolved: 
> open: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/.tmpz1HceO/lorem_ipsum.txt
> run: /Users/runner/work/coreutils/coreutils/target/debug/coreutils sum -s
> open: /var/folders/24/8k48jl6d249_n_qfxwsl6xvm0000gn/T/.tmpz1HceO/sysv_stdin.expected
> ---- test_sum::test_sysv_stdin stderr ----
> thread 'main' panicked at 'assertion failed: `(left == right)`
> Diff< left / right > :
>  [
>      54,
>      57,
>      56,
>      53,
>      32,
> <    49,
> >    50,
>      10,
>  ]
> ', tests/common/util.rs:261:9
> note: run with `RUST_BACKTRACE=1` environment variable to display a backtrace
> ```

Fixed, thanks.