@resistor 2022-07-25
> can you please document your benchmark with hyperfine in src/uu/sum/BENCHMARKING.md ?
> thanks

Done