@sylvestre 2022-07-27
On my SSD system:
```
  './target/release/coreutils sum wikidatawiki-latest-pages-logging.xml' ran
    1.42 ± 0.09 times faster than './target/release/coreutils.before sum wikidatawiki-latest-pages-logging.xml'
    2.07 ± 0.03 times faster than 'sum wikidatawiki-latest-pages-logging.xml'
```

bravo for the improvement :)